moved userid, groupid to common
