+++ src/main/scala/ironoxide/v1/common/SucceededResult.scala
@@ -1,8 +1,6 @@
 package ironoxide.v1.common
 
 import com.ironcorelabs.{sdk => jsdk}
-import ironoxide.v1.group.GroupId
-import ironoxide.v1.user.UserId
 
 /**
  * A successful edit of the access list of a document.
