+++ src/main/scala/ironoxide/v1/document/DocumentEncryptOpts.scala
@@ -3,9 +3,7 @@ package ironoxide.v1.document
 import cats.effect.Sync
 import cats.implicits._
 import com.ironcorelabs.{sdk => jsdk}
-import ironoxide.v1.common.PolicyGrant
-import ironoxide.v1.group.GroupId
-import ironoxide.v1.user.UserId
+import ironoxide.v1.common.{GroupId, PolicyGrant, UserId}
 
 /**
  * Options for encrypting a document.
