+++ src/main/scala/ironoxide/v1/group/GroupCreateOpts.scala
@@ -3,7 +3,7 @@ package ironoxide.v1.group
 import cats.effect.Sync
 import cats.implicits._
 import com.ironcorelabs.{sdk => jsdk}
-import ironoxide.v1.user.UserId
+import ironoxide.v1.common.{GroupId, UserId}
 
 /**
  * Options for group creation.
