+++ src/main/scala/ironoxide/v1/group/GroupGetResult.scala
@@ -1,8 +1,7 @@
 package ironoxide.v1.group
 
 import com.ironcorelabs.{sdk => jsdk}
-import ironoxide.v1.common.{OptionalOps, PublicKey}
-import ironoxide.v1.user.UserId
+import ironoxide.v1.common.{GroupId, OptionalOps, PublicKey, UserId}
 import java.{util => ju}
 
 /**
