+++ src/main/scala/ironoxide/v1/group/GroupMetaResult.scala
@@ -1,7 +1,7 @@
 package ironoxide.v1.group
 
 import com.ironcorelabs.{sdk => jsdk}
-import ironoxide.v1.common.OptionalOps
+import ironoxide.v1.common.{GroupId, OptionalOps}
 import java.{util => ju}
 
 /**
